+++ srcs/expansion/sepalate_str.c
@@ -6,12 +6,12 @@
 /*   By: yootaki <yootaki@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/04 13:37:28 by yootaki           #+#    #+#             */
-/*   Updated: 2021/09/05 21:22:28 by yootaki          ###   ########.fr       */
+/*   Updated: 2021/09/06 23:08:24 by yootaki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../builtin_cmd/builtin_cmd.h"
-#include "expansion.h"
+#include "../../includes/expansion.h"
 
 void	add_cmd_lst(t_cmd_lst *now)
 {
@@ -22,6 +22,15 @@ void	add_cmd_lst(t_cmd_lst *now)
 	newlst->next = now->next;
 	now->next->prev = newlst;
 	now->next = newlst;
+}
+
+/* ここの分割条件は環境変数名と同じ感じにしてもいいかも */
+int	is_sepalate_char(char c)
+{
+	if (c == ' ' || c == '\t' || c == '\0')
+		return (EXIT_FAILURE);
+	else
+		return (EXIT_SUCCESS);
 }
 
 int	sep_str(t_cmd_lst *now, t_expanser *expanser)
@@ -37,7 +46,7 @@ int	sep_str(t_cmd_lst *now, t_expanser *expanser)
 		if (expanser->str[expanser->str_cnt] == '\0')
 			return (count);
 		start = expanser->str_cnt;
-		while (expanser->str[expanser->str_cnt] != ' ' && expanser->str[expanser->str_cnt] != '\t' && expanser->str[expanser->str_cnt] != '\0')
+		while (!is_sepalate_char(expanser->str[expanser->str_cnt]))
 			expanser->str_cnt++;
 		end = expanser->str_cnt;
 		extract_str = ft_substr(expanser->str, start, end - start);
@@ -51,27 +60,4 @@ int	sep_str(t_cmd_lst *now, t_expanser *expanser)
 		}
 		count++;
 	}
-
-	// while (1)
-	// {
-	// 	while (expanser->str[expanser->str_cnt] == '$')
-	// 		expanser->str_cnt++;
-	// 	if (expanser->str[expanser->str_cnt] == '\0')
-	// 		return (count);
-	// 	start = expanser->str_cnt;
-	// 	while (expanser->str[expanser->str_cnt] != '\0' \
-	// 	&& expanser->str[expanser->str_cnt] != '$')
-	// 		expanser->str_cnt++;
-	// 	end = expanser->str_cnt;
-	// 	extract_str = ft_substr(expanser->str, start, end - start);
-	// 	if (count == 0)
-	// 		now->str = extract_str;
-	// 	else
-	// 	{
-	// 		add_cmd_lst(now);
-	// 		now = now->next;
-	// 		now->str = extract_str;
-	// 	}
-	// 	count++;
-	// }
 }
